Limits (#122)
* add in missing validation constants

* implement max autostar emoji limit

* fix emoji validation, add relative time validation

* remove todo comments

* validate max vote emojis

* validate relative time for overrides

* implement max channels per override

* add constant for max permroles

* validate max autostar channels

* implement max permroles

* implement max overrides per sb

* implement max starboards

* fix & tweak some limits

+++ src/core/starboard/vote_status.rs
@@ -81,7 +81,7 @@ impl VoteStatus {
             }
 
             // message age in seconds
-            let message_age = ((now - message_id.timestamp() as u128) / 1000) as i64;
+            let message_age = ((now as i128 - message_id.timestamp() as i128) / 1000) as i64;
 
             let min_age = config.resolved.older_than;
             let max_age = config.resolved.newer_than;
